requirements.txt
Django==4.0.2
djangorestframework
djangorestframework-simplejwt
django-filter
drf-yasg
markdown
mysql-connector-python
psycopg2-binary
python-decouple
pytest-django
